Finished create server feature
Added a create server button on account page. This button redirects the users to a form they fill out with the server info. After submitting form, it creates the server and adds the current user to the server. The user then presses a next button which transfers ownership from the bot to the user and takes the user to the details page of that server.